disable re-render
